extracted jobs logic out of reconciler, and jobs manager class added
